SISRP-25362 - Refactors server-side Degree Progress code to (1) create separate models for student and advisor and (2) prepare for the addition of new models for undergrad data (student and advisor)
