+++ app/controllers/campus_solutions/degree_progress_controller.rb
@@ -1,10 +1,9 @@
 module CampusSolutions
   class DegreeProgressController < CampusSolutionsController
-
     include AllowDelegateViewAs
 
     def get
-      render json: MyAcademics::DegreeProgress.from_session(session).get_feed_as_json
+      render json: ::DegreeProgress::MyGraduateMilestones.from_session(session).get_feed_as_json
     end
 
   end
